requirements.txt
appdirs==1.4.4
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
docker-pycreds==0.4.0
filelock==3.12.2
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.31
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.7.0
joblib==1.3.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.1
pandas==1.3.5
pathtools==0.1.2
Pillow==9.5.0
protobuf==4.23.3
psutil==5.9.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
sacremoses==0.0.53
safetensors==0.3.1
scikit-learn==1.0.2
scipy==1.7.3
sentry-sdk==1.27.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
threadpoolctl==3.1.0
tokenizers==0.13.3
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.3
wandb==0.15.4
zipp==3.15.0
